requirements.txt
flask[async]
flask-admin[async]
flask-login[async]
Flask-SQLAlchemy
Flask-Migrate
alembic
flask-jwt-extended
werkzeug
wtforms
MarkupSafe
SQLAlchemy
PyMySQL
itsdangerous
Jinja2
zipp
gunicorn
Flask-WTF
flask-seasurf
flask-talisman
mutagen
markdown
python-dotenv